Move download_url to DB.Resource
