+++ apps/transport/lib/transport_web/templates/dataset/_resource.html.eex
@@ -163,7 +163,7 @@
         <a href="<%= resource_path(@conn, :details, @resource.id) %>">
           <button class="button-outline secondary small"><i class="icon icon--plus" aria-hidden="true"></i><%= dgettext("page-dataset-details", "details") %></button>
         </a>
-        <a class="download-button" href="<%= download_url(@conn, @resource) %>">
+        <a class="download-button" href="<%= DB.Resource.download_url(@conn, @resource) %>">
           <button class="button-outline primary small"><i class="icon icon--download" aria-hidden="true"></i><%= dgettext("page-dataset-details", "Download") %></button>
         </a>
       </div>
